Made middleware for sharing logic between controller methods
