+++ routes/courses.js
@@ -7,9 +7,21 @@ const {
   deleteSingleCourse,
 } = require('../controllers/courses');
 
+const Course = require('../models/Course');
+const advanceResults = require('../middleware/advanceResults');
+
 const router = express.Router({ mergeParams: true });
 
-router.route('/').get(getCourses).post(addSingleCourse);
+router
+  .route('/')
+  .get(
+    advanceResults(Course, {
+      path: 'bootcamp',
+      select: 'name description',
+    }),
+    getCourses
+  )
+  .post(addSingleCourse);
 router
   .route('/:id')
   .get(getSingleCourse)
